backend/requirements.txt
# Google / Firebase
firebase-admin>=6.5.0
google-cloud-firestore>=2.19.0
google-cloud-secret-manager>=2.20.0

# LLM / embeddings
openai>=1.30.0

# API server
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# HTTP client (scraper, OpenRouter calls)
httpx>=0.27.0

# Utilities
tiktoken>=0.7.0
numpy>=1.26.0
schedule>=1.2.0
python-dotenv>=1.0.0
pydantic>=2.7.0
